+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed an issue where the Mythic+ timer would not properly count all minions
 - Fixed a bug that was causing buttons to appear on the chat frame
 - Fixed an issue where the chat message entries would not size properly within dungeons
+- Fixed incorrect tooltip text in the Mythic+ data panel text
 
 #### New
 
@@ -14,3 +15,4 @@
 - Introduced a setup wizard for onboarding
 - Re-skinned buttons that are available on notifications to the in-house skin
 - Revamp of global Appearances, including the ability to reset, global font, status texture, mute sound, class icon picker, etc.
+- Added a frame to the keystone frame which provides buttons to start a ready check, start a pull timer, and enable or disable automatically starting the key after a pull timer.